Add map_binary_state_to_semantic utility function and update entity logic
- Introduced a new utility function `map_binary_state_to_semantic` to convert binary sensor states ('on'/'off') to semantic states ('open'/'closed') based on active states.
- Updated the `Entity` class to utilize the new mapping function for state validation.
- Removed the old mapping function from the correlation module and adjusted related logic to ensure consistency across the codebase.
- Added unit tests for the new mapping function to verify its correctness and behavior with various input scenarios.
