+++ tests/test_utils.py
@@ -12,6 +12,7 @@ from custom_components.area_occupancy.utils import (
     combine_priors,
     format_float,
     format_percentage,
+    map_binary_state_to_semantic,
 )
 
 
@@ -649,3 +650,46 @@ class TestBayesianProbability:
         entity.get_likelihoods.assert_called_once()
         # Result should be based on static prob_given_true/prob_given_false
         assert result > 0.5
+
+
+class TestMapBinaryStateToSemantic:
+    """Test map_binary_state_to_semantic function.
+
+    Tests mapping of binary sensor states ('on'/'off') to semantic states
+    ('open'/'closed') for door and window sensors.
+    """
+
+    @pytest.mark.parametrize(
+        ("input_state", "active_states", "expected_result", "description"),
+        [
+            ("off", ["closed"], "closed", "door closed (off -> closed)"),
+            ("on", ["open"], "open", "door open (on -> open)"),
+            ("on", ["open"], "open", "window open (on -> open)"),
+            ("off", ["closed"], "closed", "window closed (off -> closed)"),
+        ],
+    )
+    def test_map_binary_state_to_semantic(
+        self, input_state, active_states, expected_result, description
+    ):
+        """Test mapping binary states to semantic states."""
+        result = map_binary_state_to_semantic(input_state, active_states)
+        assert result == expected_result
+
+    @pytest.mark.parametrize(
+        ("input_state", "active_states", "expected_result"),
+        [
+            ("off", ["on"], "off"),  # No mapping when semantic not in active_states
+            ("on", ["off"], "on"),  # No mapping when semantic not in active_states
+        ],
+    )
+    def test_no_mapping_when_semantic_not_present(
+        self, input_state, active_states, expected_result
+    ):
+        """Test that no mapping occurs when semantic states not in active_states."""
+        result = map_binary_state_to_semantic(input_state, active_states)
+        assert result == expected_result
+
+    def test_mapping_preserves_other_states(self):
+        """Test that non-binary states are preserved."""
+        result = map_binary_state_to_semantic("playing", ["playing", "paused"])
+        assert result == "playing"
